+++ src/components/Rate/index.ts
@@ -1,3 +1,0 @@
-import PRate from './Rate.vue';
-
-export default PRate;