+++ CHANGELOG.md
@@ -1,3 +1,58 @@
+# [0.13.0](https://github.com/verteva/photon/compare/v0.12.0...v0.13.0) (2022-12-12)
+
+
+### Bug Fixes
+
+* move from rem to px ([73530d4](https://github.com/verteva/photon/commit/73530d408750ce9cafe8e7ee35c7e231031a6f62))
+* unit tests ([7ae7539](https://github.com/verteva/photon/commit/7ae75398a39754ffb2ddd1f102337e6e29b5e01e))
+* unit tests ([f35abf9](https://github.com/verteva/photon/commit/f35abf9ffceccf4122146c7b502fa35d51e6b83e))
+* update medium from from using 'med -> md' ([b5e8edb](https://github.com/verteva/photon/commit/b5e8edbab6e2736a7ac8cd340bbadca28ec99264))
+* wip CX-31 input text update ([73616cf](https://github.com/verteva/photon/commit/73616cf4b341280cd639585d98a03ab73c81cee3))
+
+
+### Features
+
+* skip link v2 ([1fa43c1](https://github.com/verteva/photon/commit/1fa43c10908924aede94960536f0d14b566322ed))
+* skiplink v2 with adjustments ([cceef3e](https://github.com/verteva/photon/commit/cceef3e66a3e7965a7b29b23aad68d905cb0d67f))
+
+# [0.12.0](https://github.com/verteva/photon/compare/v0.11.0...v0.12.0) (2022-12-05)
+
+
+### Bug Fixes
+
+* add info card photon v2 to export ob-web ([ca7fe0b](https://github.com/verteva/photon/commit/ca7fe0b4ae708b97a99f9300d8101cb0ffd1c23d))
+* incorrect label variable props defined in atoms component ([dff030a](https://github.com/verteva/photon/commit/dff030ac8e448e38f7f5c5f3f9991746fd497ac4))
+* re-arrange css values ([6004a42](https://github.com/verteva/photon/commit/6004a425aa3c0b3273c55530148acd3190d08250))
+* update export names of list items ([d52af6b](https://github.com/verteva/photon/commit/d52af6b3d49b7c3cb7d21c1faf87d2d0d7ec2589))
+* update indicator and position of dropdown ([a216f78](https://github.com/verteva/photon/commit/a216f786e7ec51cfdaf4359492e44dab32eb24eb))
+* update input select found issues during integration ([0b61e01](https://github.com/verteva/photon/commit/0b61e018bbf964ea9a1a1b0ff27267b8e11ce8f9))
+
+
+### Features
+
+* brand gradient text photon v2 components ([e38ab89](https://github.com/verteva/photon/commit/e38ab8939e2f2f6bea404a6f698cf5d6f1d2eeaf))
+
+# [0.11.0](https://github.com/verteva/photon/compare/v0.10.0...v0.11.0) (2022-11-30)
+
+
+### Bug Fixes
+
+* change to relative path ([b161593](https://github.com/verteva/photon/commit/b16159367dd8d6f7c5d2e8dcda9afec980e99c21))
+* changed coverage provider ([e1ba3f3](https://github.com/verteva/photon/commit/e1ba3f3988725c45d3a6daa3ddf1bb9745101377))
+* missing footer style and filter binding ([6e30ad1](https://github.com/verteva/photon/commit/6e30ad11e658ac1dfd129e0e7dac53dfc3a8cd48))
+* move props outside ([c2d0aa3](https://github.com/verteva/photon/commit/c2d0aa3d4bb0443f829baade0cfba07a4102da3f))
+* removed react deps ([6a235af](https://github.com/verteva/photon/commit/6a235af3e7dfc4d96715579b7d435f30cf7a7b3f))
+* update name of component test ([3c35e1f](https://github.com/verteva/photon/commit/3c35e1f443a3468aaced291a4724f91330c1b954))
+
+
+### Features
+
+* accordion v2 component ([b9d9524](https://github.com/verteva/photon/commit/b9d9524fe43cc6b7e9c7d03d8d2391d61e87f2b4))
+* content card v2 ([7d9712f](https://github.com/verteva/photon/commit/7d9712f4523f21026320e260d6fe186083d27522))
+* info card photon version 2 component ([644cc17](https://github.com/verteva/photon/commit/644cc17b2a163c8216b11b7fbdb01f0cd9d3ca64))
+* list items component photon V2 ([01ee3c1](https://github.com/verteva/photon/commit/01ee3c1536a61dc12896ad7811c2e3c1c23fe328))
+* photon v2 input select integration ([a2c8f7b](https://github.com/verteva/photon/commit/a2c8f7b198d8ed88ed1a86dfa022fd962812adc5))
+
 # [0.10.0](https://github.com/verteva/photon/compare/v0.9.0...v0.10.0) (2022-11-27)